pyproject.toml
[tool.poetry]
name = "tier-synthesis"
version = "0.1.0"
description = "A collaborative tier list platform for analyzing character preferences"
authors = ["Taksqth <gabriel.takeshi.higa@gmail.com>"]
readme = "README.md"
license = "MIT"
keywords = ["tier-list", "collaborative-filtering", "character-analysis"]

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.12.0"
ipykernel = "^6.29.5"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"